init commit
I've got the email provider working! There are no real database calls yet, but they should work.

The main problem is that there seems to be no way to use a middleware.ts file... If one is present, The console fills with errors and the site crashes. Once removed/renamed, everything works as expected. There are a couple of resources to look back into.

The tutorials I'm following don't seem to have this issue. One uses mongodb, another uses prisma/neon. The prisma one mentions how the database doesn't support the edge and how to work around it, but even they never hit the errors. The moment they added the middleware, they were able to log stuff. I can't even get that far.

The console's errors are filled with 'mongodb' references... but surely there's a way to get this to work. Maybe the issue is with the beta version of next-auth I'm using?

Also, look into using the signIn/signOut function exported from the @/auth.ts file. Use those when possible instead of the ones imported from next-auth/react.